wip: chat rewrite
